Specify wan by site name, not pod name
Later we can add the functionality to specify one of
multiple wan sources at a site.
But for now, let's just specify the site to get WAN from, because
hardware.yaml currently only allows for one wan_bridge value.

This fixes a bug that happens when the wan bridge's name
varies across pods in a site.

Signed-off-by: Andy Strohman <andrew@andrewstrohman.com>